DRY up `render` command tests
